row: move tests to a separate file
If only tests are modified when put in the same file as library code,
Cargo will rebuild the lib crate anyway. Conversely, when tests are in
a separate file, the lib crate won't be rebuilt and this saves precious
time.
